# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ConclusionItem": {
        "properties": {
          "caveats": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Caveats",
            "type": "array"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "disposition": {
            "title": "Disposition",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "key_evidence_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Key Evidence Ids",
            "type": "array"
          },
          "statement": {
            "title": "Statement",
            "type": "string"
          },
          "supporting_hypothesis_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Supporting Hypothesis Ids",
            "type": "array"
          }
        },
        "required": [
          "id",
          "statement",
          "disposition",
          "confidence"
        ],
        "title": "ConclusionItem",
        "type": "object"
      },
...
      "CritiqueItem": {
        "properties": {
          "critique_type": {
            "title": "Critique Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "resolved": {
            "default": false,
            "title": "Resolved",
            "type": "boolean"
          },
          "severity": {
            "title": "Severity",
            "type": "string"
          },
          "suggested_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Action"
          },
          "target_id": {
            "title": "Target Id",
            "type": "string"
          },
          "target_kind": {
            "title": "Target Kind",
            "type": "string"
          }
        },
        "required": [
          "id",
          "critique_type",
          "severity",
          "target_kind",
          "target_id",
          "message"
        ],
        "title": "CritiqueItem",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "InvestigationCounts": {
        "properties": {
          "critiques": {
            "default": 0,
            "title": "Critiques",
            "type": "integer"
          },
          "decisions": {
            "default": 0,
            "title": "Decisions",
            "type": "integer"
          },
          "evidence": {
            "default": 0,
            "title": "Evidence",
            "type": "integer"
          },
          "experiments": {
            "default": 0,
            "title": "Experiments",
            "type": "integer"
          },
          "hypotheses": {
            "default": 0,
            "title": "Hypotheses",
            "type": "integer"
          },
          "observations": {
            "default": 0,
            "title": "Observations",
            "type": "integer"
          },
          "open_questions": {
            "default": 0,
            "title": "Open Questions",
            "type": "integer"
          }
        },
        "title": "InvestigationCounts",
        "type": "object"
      },
      "InvestigationCreateRequest": {
        "properties": {
          "async_execution": {
            "default": false,
            "description": "Run in the background via the worker (requires a running worker) instead of synchronously.",
            "title": "Async Execution",
            "type": "boolean"
          },
          "dataset": {
            "$ref": "#/components/schemas/InvestigationDatasetInput"
          },
          "goal": {
            "description": "What the investigation should answer.",
            "minLength": 1,
            "title": "Goal",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "goal",
          "dataset"
        ],
        "title": "InvestigationCreateRequest",
        "type": "object"
      },
      "InvestigationCreateResponse": {
        "properties": {
          "analysis_run_id": {
            "format": "uuid",
            "title": "Analysis Run Id",
            "type": "string"
          },
          "db_status": {
            "title": "Db Status",
            "type": "string"
          },
          "investigation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigation Id"
          },
          "queued": {
            "default": false,
            "title": "Queued",
            "type": "boolean"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "analysis_run_id",
          "status",
          "db_status"
        ],
        "title": "InvestigationCreateResponse",
        "type": "object"
      },
      "InvestigationDatasetInput": {
        "description": "What an investigation should run over.\n\n``source`` selects the adapter. It defaults to ``tabular`` so every caller written before\nEDGAR was reachable here keeps working unchanged \u2014 the field is additive, not a new\nrequired discriminator.\n\nThe fields are flat rather than a tagged union because the two sources overlap (both carry\na name and a goal-facing identity) and because a union would have broken that\ncompatibility. ``check_source_fields`` enforces which fields each source actually needs, so\nthe looseness is in the shape, not in the validation.",
        "properties": {
          "csv_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Csv Text"
          },
          "entities": {
            "description": "Ticker symbols when source is 'edgar' (e.g. ['AAPL', 'MSFT']).",
            "items": {
              "type": "string"
            },
            "title": "Entities",
            "type": "array"
          },
          "entity_id_fields": {
            "items": {
              "type": "string"
            },
            "title": "Entity Id Fields",
            "type": "array"
          },
          "format": {
            "default": "csv",
            "enum": [
              "csv",
              "records"
            ],
            "title": "Format",
            "type": "string"
          },
          "name": {
            "default": "dataset",
            "title": "Name",
            "type": "string"
          },
          "records": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Records"
          },
          "refresh": {
            "default": false,
            "description": "Force a fresh SEC fetch rather than reusing cached filings.",
            "title": "Refresh",
            "type": "boolean"
          },
          "source": {
            "default": "tabular",
            "enum": [
              "tabular",
              "edgar"
            ],
            "title": "Source",
            "type": "string"
          },
          "time_field": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Field"
          }
        },
        "title": "InvestigationDatasetInput",
        "type": "object"
      },
      "InvestigationDetail": {
        "properties": {
          "adapter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Id"
          },
          "analysis_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Run Id"
          },
          "conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conclusion"
          },
          "conclusion_detail": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ConclusionItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "constraints": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Constraints",
            "type": "array"
          },
          "counts": {
            "$ref": "#/components/schemas/InvestigationCounts"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "critiques": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/CritiqueItem"
            },
            "title": "Critiques",
            "type": "array"
          },
          "datasets": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DatasetItem"
            },
            "title": "Datasets",
            "type": "array"
          },
          "decisions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/DecisionItem"
            },
            "title": "Decisions",
            "type": "array"
          },
          "domain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "events": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EventItem"
            },
            "title": "Events",
            "type": "array"
          },
          "evidence": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/EvidenceItem"
            },
            "title": "Evidence",
            "type": "array"
          },
          "experiments": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ExperimentItem"
            },
            "title": "Experiments",
            "type": "array"
          },
          "hypotheses": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/HypothesisItem"
            },
            "title": "Hypotheses",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "observations": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ObservationItem"
            },
            "title": "Observations",
            "type": "array"
          },
          "open_questions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/OpenQuestionItem"
            },
            "title": "Open Questions",
            "type": "array"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "success_criteria": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Success Criteria",
            "type": "array"
          },
          "termination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TerminationView"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin",
          "status",
          "counts",
          "created_at",
          "updated_at"
        ],
        "title": "InvestigationDetail",
        "type": "object"
      },
      "InvestigationSummary": {
        "properties": {
          "adapter_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Adapter Id"
          },
          "analysis_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Run Id"
          },
          "conclusion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conclusion"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "counts": {
            "$ref": "#/components/schemas/InvestigationCounts"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "objective": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objective"
          },
          "origin": {
            "title": "Origin",
            "type": "string"
          },
          "project_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "origin",
          "status",
          "counts",
          "created_at",
          "updated_at"
        ],
        "title": "InvestigationSummary",
        "type": "object"
      },
...
      "ReplayResponse": {
        "properties": {
          "analysis_run_id": {
            "title": "Analysis Run Id",
            "type": "string"
          },
          "candidate_persisted": {
            "default": false,
            "description": "Always false. The candidate is deliberately not stored \u2014 sharing a store with the baseline would overwrite the very run being compared against.",
            "title": "Candidate Persisted",
            "type": "boolean"
          },
          "diff": {
            "$ref": "#/components/schemas/ReplayDiffView"
          },
          "investigation_id": {
            "title": "Investigation Id",
            "type": "string"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "title": "Notes",
            "type": "array"
          },
          "same_dataset": {
            "description": "False when the candidate ran against data other than the baseline's, in which case a divergence cannot be attributed to the policy change alone.",
            "title": "Same Dataset",
            "type": "boolean"
          }
        },
        "required": [
          "investigation_id",
          "analysis_run_id",
          "diff",
          "same_dataset"
        ],
        "title": "ReplayResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "EDGAR Analytics API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/investigations": {
      "get": {
        "description": "List investigations owned by the user's projects (or initiated by the user), newest first.",
        "operationId": "list_investigations_v1_investigations_get",
        "parameters": [
          {
            "description": "Scope to one owned project",
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Scope to one owned project",
              "title": "Project Id"
            }
          },
          {
            "description": "Resolve the investigation for a run",
            "in": "query",
            "name": "analysis_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Resolve the investigation for a run",
              "title": "Analysis Run Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InvestigationSummary"
                  },
                  "title": "Response List Investigations V1 Investigations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Investigations",
        "tags": [
          "investigations"
        ]
      },
      "post": {
        "description": "Create an agentic investigation over a dataset and run it.\n\n``dataset.source`` selects the adapter: ``tabular`` (pasted CSV or inline records, the\ndefault) or ``edgar`` (SEC filings for ``dataset.entities``). Both run the same loop over\nthe same evidence model \u2014 EDGAR is an adapter, not a separate product.\n\nSynchronous by default; ``async_execution=true`` enqueues it for the worker. Prefer async\nfor ``edgar``: the panel is built from live SEC fetches before the loop starts, which is\nfar too slow to hold an HTTP request open for.\n\nRequires the agentic engine flag; owner-scoped to the target project.",
        "operationId": "create_investigation_v1_investigations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestigationCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationCreateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Investigation",
        "tags": [
          "investigations"
        ]
      }
    },
    "/v1/investigations/{investigation_id}": {
      "get": {
        "description": "Full investigation: hypotheses, evidence, agent decisions, critiques, experiments, conclusion, timeline.",
        "operationId": "get_investigation_v1_investigations__investigation_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InvestigationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Investigation",
        "tags": [
          "investigations"
        ]
      }
    },
    "/v1/investigations/{investigation_id}/replay": {
      "post": {
        "description": "Re-run this investigation under the currently configured policy and diff the outcome.\n\nAnswers \"did changing the model / prompt / budget change our answers, or only the route to\nthem?\" against a real persisted run. The verdict is conclusion-first: ``identical``,\n``same_conclusion``, or ``diverged``.\n\n**Admin only, and deliberately so.** A replay is a full investigation and costs real model\nspend, but the candidate is never persisted \u2014 that is a design decision in\n:mod:`agentic.agent.replay`, since sharing a store with the baseline would overwrite the\nrun being compared against. The consequence is that a replay produces no ``AnalysisRun``\nand no ``ModelCall`` rows, so :mod:`backend.services.spend_guard` cannot see or count it.\nAn endpoint that spends money invisibly to the ceilings must not be reachable by ordinary\naccounts.\n\n**Synchronous**, and as slow as the investigation it replays \u2014 tens of seconds against a\nreal model. There is no background variant: the candidate is not persisted, so there would\nbe nothing for a worker to hand back.\n\nRefuses with **409** when the baseline's dataset can no longer be reconstructed. Replay\nreuses the exact panel the baseline ran on; re-fetching would change the data underneath\nthe comparison and attribute to the policy a difference that came from the data.",
        "operationId": "replay_investigation_route_v1_investigations__investigation_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "investigation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Investigation Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replay Investigation Route",
        "tags": [
          "investigations"
        ]
      }
    },
...
  }
}